# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Flow — Safe to Spend API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface IncomeEntry {
  id: number;
  amount: number;
  source: string;
  date: string;
  category: string;
}

export interface IncomeInput {
  amount: number;
  source?: string;
  date: string;
  category?: string;
}

export interface Bill {
  id: number;
  amount: number;
  name: string;
  category: string;
}

export interface BillInput {
  amount: number;
  name: string;
  category?: string;
}

export interface Settings {
  bufferPct: number;
  bufferBalance: number;
  bufferGoalMonths: number;
  onboarded: boolean;
}

export interface SettingsUpdate {
  bufferPct?: number;
  bufferBalance?: number;
  bufferGoalMonths?: number;
  onboarded?: boolean;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  income: 'income',
  bill: 'bill',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  label: string;
  amount: number;
  /** @nullable */
  date: string | null;
  category: string;
}

export type DashboardInsightsBudgetHealth = typeof DashboardInsightsBudgetHealth[keyof typeof DashboardInsightsBudgetHealth];


export const DashboardInsightsBudgetHealth = {
  excellent: 'excellent',
  good: 'good',
  fair: 'fair',
  tight: 'tight',
} as const;

export interface DashboardInsights {
  dailyBudget: number;
  daysRemainingInMonth: number;
  savingsPct: number;
  spendingRate: number;
  /** Projected balance at end of month (safeToSpend - dailyBudget*daysRemaining placeholder) */
  projectedMonthEnd: number;
  budgetHealth: DashboardInsightsBudgetHealth;
}

export interface Dashboard {
  safeToSpend: number;
  baselineIncome: number;
  totalBills: number;
  bufferBalance: number;
  bufferGoal: number;
  bufferPct: number;
  bufferGoalMonths: number;
  onboarded: boolean;
  recentActivity: ActivityItem[];
  insights: DashboardInsights;
}

export interface MonthlyData {
  month: number;
  year: number;
  label: string;
  income: number;
  bills: number;
}

export type RecurringItemType = typeof RecurringItemType[keyof typeof RecurringItemType];


export const RecurringItemType = {
  income: 'income',
  bill: 'bill',
} as const;

export type RecurringItemFrequency = typeof RecurringItemFrequency[keyof typeof RecurringItemFrequency];


export const RecurringItemFrequency = {
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
  yearly: 'yearly',
} as const;

export interface RecurringItem {
  id: number;
  type: RecurringItemType;
  name: string;
  amount: number;
  category: string;
  frequency: RecurringItemFrequency;
  startDate: string;
  active: boolean;
}

export type RecurringInputType = typeof RecurringInputType[keyof typeof RecurringInputType];


export const RecurringInputType = {
  income: 'income',
  bill: 'bill',
} as const;

export type RecurringInputFrequency = typeof RecurringInputFrequency[keyof typeof RecurringInputFrequency];


export const RecurringInputFrequency = {
  daily: 'daily',
  weekly: 'weekly',
  monthly: 'monthly',
  yearly: 'yearly',
} as const;

export interface RecurringInput {
  type: RecurringInputType;
  name: string;
  amount: number;
  category?: string;
  frequency: RecurringInputFrequency;
  startDate: string;
  active?: boolean;
}

export interface SavingsGoal {
  id: number;
  name: string;
  targetAmount: number;
  currentAmount: number;
  /** @nullable */
  deadline?: string | null;
  completed: boolean;
  createdAt?: string;
}

export interface SavingsGoalInput {
  name: string;
  targetAmount: number;
  currentAmount?: number;
  deadline?: string;
  completed?: boolean;
}

export interface SavingsGoalUpdate {
  name?: string;
  targetAmount?: number;
  currentAmount?: number;
  deadline?: string;
  completed?: boolean;
}

export interface DemoResult {
  message: string;
}

export type ListIncomeParams = {
search?: string;
category?: string;
dateFrom?: string;
dateTo?: string;
amountMin?: number;
amountMax?: number;
};

export type ListBillsParams = {
search?: string;
category?: string;
amountMin?: number;
amountMax?: number;
};

export type GetMonthlyAnalyticsParams = {
months?: number;
};

export type ExportCsvParams = {
type: ExportCsvType;
};

export type ExportCsvType = typeof ExportCsvType[keyof typeof ExportCsvType];


export const ExportCsvType = {
  income: 'income',
  bills: 'bills',
  all: 'all',
} as const;

export type ListRecurringParams = {
type?: ListRecurringType;
};

export type ListRecurringType = typeof ListRecurringType[keyof typeof ListRecurringType];


export const ListRecurringType = {
  income: 'income',
  bill: 'bill',
} as const;