# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Flow — Safe to Spend API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: income
  - name: bills
  - name: settings
  - name: dashboard
  - name: analytics
  - name: export
  - name: recurring
  - name: savings-goals
  - name: demo
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /income:
    get:
      operationId: listIncome
      tags: [income]
      summary: List income entries
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: amountMin
          in: query
          schema:
            type: number
        - name: amountMax
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of income entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IncomeEntry"
    post:
      operationId: createIncome
      tags: [income]
      summary: Create an income entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeInput"
      responses:
        "201":
          description: Created income entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeEntry"
        "400":
          description: Invalid input

  /income/{id}:
    patch:
      operationId: updateIncome
      tags: [income]
      summary: Update an income entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncomeInput"
      responses:
        "200":
          description: Updated income entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeEntry"
        "404":
          description: Not found
    delete:
      operationId: deleteIncome
      tags: [income]
      summary: Delete an income entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /bills:
    get:
      operationId: listBills
      tags: [bills]
      summary: List fixed bills
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: amountMin
          in: query
          schema:
            type: number
        - name: amountMax
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
    post:
      operationId: createBill
      tags: [bills]
      summary: Create a fixed bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillInput"
      responses:
        "201":
          description: Created bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "400":
          description: Invalid input

  /bills/{id}:
    patch:
      operationId: updateBill
      tags: [bills]
      summary: Update a fixed bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BillInput"
      responses:
        "200":
          description: Updated bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "404":
          description: Not found
    delete:
      operationId: deleteBill
      tags: [bills]
      summary: Delete a fixed bill
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get user settings
      responses:
        "200":
          description: User settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Aggregated dashboard summary with insights
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"

  /analytics/monthly:
    get:
      operationId: getMonthlyAnalytics
      tags: [analytics]
      summary: Monthly income and bills data for charts
      parameters:
        - name: months
          in: query
          schema:
            type: integer
            default: 6
      responses:
        "200":
          description: Monthly analytics data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyData"

  /export/csv:
    get:
      operationId: exportCsv
      tags: [export]
      summary: Export all data as CSV
      parameters:
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [income, bills, all]
      responses:
        "200":
          description: CSV data
          content:
            text/csv:
              schema:
                type: string

  /recurring:
    get:
      operationId: listRecurring
      tags: [recurring]
      summary: List recurring items
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, bill]
      responses:
        "200":
          description: List of recurring items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecurringItem"
    post:
      operationId: createRecurring
      tags: [recurring]
      summary: Create a recurring item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurringInput"
      responses:
        "201":
          description: Created recurring item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringItem"

  /recurring/{id}:
    put:
      operationId: updateRecurring
      tags: [recurring]
      summary: Update a recurring item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurringInput"
      responses:
        "200":
          description: Updated recurring item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecurringItem"
    delete:
      operationId: deleteRecurring
      tags: [recurring]
      summary: Delete a recurring item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /savings-goals:
    get:
      operationId: listSavingsGoals
      tags: [savings-goals]
      summary: List savings goals
      responses:
        "200":
          description: List of savings goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SavingsGoal"
    post:
      operationId: createSavingsGoal
      tags: [savings-goals]
      summary: Create a savings goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavingsGoalInput"
      responses:
        "201":
          description: Created savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"

  /savings-goals/{id}:
    put:
      operationId: updateSavingsGoal
      tags: [savings-goals]
      summary: Update a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SavingsGoalUpdate"
      responses:
        "200":
          description: Updated savings goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SavingsGoal"
    delete:
      operationId: deleteSavingsGoal
      tags: [savings-goals]
      summary: Delete a savings goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /demo/load:
    post:
      operationId: loadDemoData
      tags: [demo]
      summary: Load demo data (clears and seeds)
      responses:
        "200":
          description: Demo data loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoResult"

  /demo/clear:
    post:
      operationId: clearAllData
      tags: [demo]
      summary: Clear all user data
      responses:
        "200":
          description: All data cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoResult"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    IncomeEntry:
      type: object
      required: [id, amount, source, date, category]
      properties:
        id:
          type: integer
        amount:
          type: number
        source:
          type: string
        date:
          type: string
          format: date
        category:
          type: string

    IncomeInput:
      type: object
      required: [amount, date]
      properties:
        amount:
          type: number
        source:
          type: string
        date:
          type: string
          format: date
        category:
          type: string

    Bill:
      type: object
      required: [id, amount, name, category]
      properties:
        id:
          type: integer
        amount:
          type: number
        name:
          type: string
        category:
          type: string

    BillInput:
      type: object
      required: [amount, name]
      properties:
        amount:
          type: number
        name:
          type: string
        category:
          type: string

    Settings:
      type: object
      required: [bufferPct, bufferBalance, bufferGoalMonths, onboarded]
      properties:
        bufferPct:
          type: number
        bufferBalance:
          type: number
        bufferGoalMonths:
          type: number
        onboarded:
          type: boolean

    SettingsUpdate:
      type: object
      properties:
        bufferPct:
          type: number
        bufferBalance:
          type: number
        bufferGoalMonths:
          type: number
        onboarded:
          type: boolean

    ActivityItem:
      type: object
      required: [id, type, label, amount, date, category]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [income, bill]
        label:
          type: string
        amount:
          type: number
        date:
          type: ["string", "null"]
        category:
          type: string

    DashboardInsights:
      type: object
      required: [dailyBudget, daysRemainingInMonth, savingsPct, spendingRate, projectedMonthEnd, budgetHealth]
      properties:
        dailyBudget:
          type: number
        daysRemainingInMonth:
          type: integer
        savingsPct:
          type: number
        spendingRate:
          type: number
        projectedMonthEnd:
          type: number
          description: Projected balance at end of month (safeToSpend - dailyBudget*daysRemaining placeholder)
        budgetHealth:
          type: string
          enum: [excellent, good, fair, tight]

    Dashboard:
      type: object
      required:
        - safeToSpend
        - baselineIncome
        - totalBills
        - bufferBalance
        - bufferGoal
        - bufferPct
        - bufferGoalMonths
        - onboarded
        - recentActivity
        - insights
      properties:
        safeToSpend:
          type: number
        baselineIncome:
          type: number
        totalBills:
          type: number
        bufferBalance:
          type: number
        bufferGoal:
          type: number
        bufferPct:
          type: number
        bufferGoalMonths:
          type: number
        onboarded:
          type: boolean
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
        insights:
          $ref: "#/components/schemas/DashboardInsights"

    MonthlyData:
      type: object
      required: [month, year, label, income, bills]
      properties:
        month:
          type: integer
        year:
          type: integer
        label:
          type: string
        income:
          type: number
        bills:
          type: number

    RecurringItem:
      type: object
      required: [id, type, name, amount, category, frequency, startDate, active]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [income, bill]
        name:
          type: string
        amount:
          type: number
        category:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, monthly, yearly]
        startDate:
          type: string
          format: date
        active:
          type: boolean

    RecurringInput:
      type: object
      required: [type, name, amount, frequency, startDate]
      properties:
        type:
          type: string
          enum: [income, bill]
        name:
          type: string
        amount:
          type: number
        category:
          type: string
        frequency:
          type: string
          enum: [daily, weekly, monthly, yearly]
        startDate:
          type: string
          format: date
        active:
          type: boolean

    SavingsGoal:
      type: object
      required: [id, name, targetAmount, currentAmount, completed]
      properties:
        id:
          type: integer
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: ["string", "null"]
          format: date
        completed:
          type: boolean
        createdAt:
          type: string

    SavingsGoalInput:
      type: object
      required: [name, targetAmount]
      properties:
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: string
          format: date
        completed:
          type: boolean

    SavingsGoalUpdate:
      type: object
      properties:
        name:
          type: string
        targetAmount:
          type: number
        currentAmount:
          type: number
        deadline:
          type: string
          format: date
        completed:
          type: boolean

    DemoResult:
      type: object
      required: [message]
      properties:
        message:
          type: string