# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Flow — Safe to Spend API
 * OpenAPI spec version: 0.2.0
 */
import type { ActivityItemType } from './activityItemType';

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  label: string;
  amount: number;
  /** @nullable */
  date: string | null;
  category: string;
}