# lib/api-zod/src/generated/types/bill.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Flow — Safe to Spend API
 * OpenAPI spec version: 0.2.0
 */

export interface Bill {
  id: number;
  amount: number;
  name: string;
  category: string;
}